tb#quest: Add patch to make it work on Linux
Ideally, all the .bat and .sh scripts should be removed and replaced with
HANDLE_TILESETS.
But the BW Fixpack doesn't yet seem to have a mechanism for sharing the
`tisunpack` binaries between mods.

With this patch, Linux and Mac users can at least install the mod correctly
if they have placed `tisunpack` in their `$PATH`.