spell_rev: Fix patch encoding
This patch file was encoded in UTF-16, which caused the `patch` utility on
Linux to reject it. The debug output it printed, was:

    ! patch -p0 -f --verbose --no-backup-if-mismatch --ignore-whitespace < 'BiG_World_Fixpack/spell_rev/main_component.tpa.patch' | tee -a '_ApplyPatches_verbose.log'
    Hmm...  I can't seem to find a patch in there anywhere.
    patch: **** Only garbage was found in the patch input.

I uses a shell commans to check the encodings of all 1759 patch files
contained in this Fixpack, and this was the *only* one that was encoded in
UTF-16, so I don't think this is something that `_ApplyPatches.tp2` needs
to be taught to handle.

So I just re-encoded the patch file in ASCII, and now it applies cleanly.