stratagems: Fix mangled patch
Something must have gone very wrong when this patch file was created, because
instead of listing the two sections for the two target files one below the
other, the section for the `/bg1/` file *replaced* thebeginning of the section
for the `/bg2/` file.

The result was a broken patch which the `patch` utility rejected.

I reconstructed the patch from the original files; now it applies cleanly.